+++ ChangeLog.md
@@ -9,6 +9,7 @@
 - [Breaking] This version cannot guarantee compatible with persisted database models file by previous version.
 
 Blob:
+- Fixed a bug that snapshot blob doesn't honor metadata options.
 - Updated Azure Storage API version to 2019-02-02, and added following new features:
   - Supports new SAS format with blob snapshot.
   - Responses now includes x-ms-client-request-id when client request ID provided in request.